Fonctionnalité graphique format log v1
